requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
mysql-connector==2.2.9 
mysqlclient==2.0.3     
SQLAlchemy==1.4.20